// src/app/modules/models/UserInfo.ts
export interface UserInfo {
    abbreviation: string;
    created_date: string;
    document_number: number;
    document_type_id: number;
    email: string;
    first_lastname: string;
    first_name: string;
    password: string;
    phone: string;
    profile_id: number;
    profile_name: string;
    second_lastname: string;
    second_name: string;
    status: string;
    user_id: number;
}